feat(shell): enhetligt app-skal — kontextuell sidebar + AppShell
Fas 1 (skal-enhetlighet), bygg-delen. Nya delar för ETT skal över hela appen:

- app/_lib/appNav.ts — rolldriven app-nivå-nav (Start + montörs/sälj/admin-
  destinationer från gamla headern/dashboarden)
- app/components/AppSidebar.tsx — kontextuell sidebar: app-nav utanför /crm,
  CRM-nav + "Till start" inne i /crm. Återanvänder CrmSidebars collapse/drawer-
  mönster, ingen global-header-offset (fullhöjd + safe-area i mobil-topbaren).
  Footern återanvänder ProfileMenu (profil/byt lösenord/logga ut).
- app/components/AppShell.tsx — klient-skal: publika sidor (/auth, /kund/offert)
  renderar bart, övrigt får sidebar + content-area. crm-shell-klassen bär temat.
- ProfileMenu: nytt placement="up" så kontomenyn öppnas uppåt i footern.
