--- app/components/HeaderTitle.tsx
@@ -1,23 +1,0 @@
-"use client";
-import { usePathname } from "next/navigation";
-
-export default function HeaderTitle() {
-  const pathname = usePathname();
-  let title = '';
-  if (pathname === '/') title = 'Startsida';
-  else if (pathname.startsWith('/egenkontroll')) title = 'Skapa egenkontroll';
-  else if (pathname.startsWith('/archive')) title = 'Sparade egenkontroller';
-  else if (pathname.startsWith('/kontakt-lista')) title = 'Kontakt & Adresser';
-  else if (pathname.startsWith('/dokument-information')) title = 'Dokument & Information';
-  else if (pathname.startsWith('/nyheter')) title = 'Nyheter';
-  else if (pathname.startsWith('/bestallning-klader')) title = 'Beställa kläder';
-  else if (pathname.startsWith('/crm/korjournal')) title = 'Körjournal';
-  else if (pathname.startsWith('/plannering')) title = 'Planering';
-  else if (pathname.startsWith('/auth/sign-in')) title = 'Logga in';
-  else if (pathname.startsWith('/auth/create-account')) title = 'Skapa konto';
-  else if (pathname.startsWith('/offert')) title = 'Skapa offert';
-  else if (pathname.startsWith('/admin')) title = 'Admin';
-  else if(pathname.startsWith('/tidrapport')) title = 'Tidrapportering';
-  else if(pathname.startsWith('/kund/offert')) title = 'Beställningsbekräftelse';
-  return <strong style={{fontSize: '14px', letterSpacing: 0.2, marginLeft: 16, marginTop: 4}}>{title}</strong>;
-}
